--- client/src/components/Global/Form/TextField.js
@@ -1,206 +1,0 @@
-import React, { useState } from "react";
-import styled, { css } from "styled-components";
-
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faEye, faEyeSlash } from "@fortawesome/free-solid-svg-icons";
-
-const TextField = ({
-  className,
-  name = "",
-  type = "text",
-  value = "",
-  label = "",
-  required = false,
-  placeholder = "",
-  infos,
-  errors,
-  validation = false,
-  onChange,
-  onFocusOut,
-  togglePassword = false,
-  inline = false,
-  suffix = false,
-}) => {
-  const [revealed, setRevealed] = useState(false);
-  const togglable = type === "password" && togglePassword;
-  return (
-    <div
-      className={`${className} ${
-        validation && errors && errors.length > 0 ? "error" : ""
-      }`}
-      data-nrt={`field-${name}`}
-    >
-      <div className={`field-wrapper ${inline ? "inline" : ""}`}>
-        <div className="label-wrapper">
-          {label && <label htmlFor={name}>{label}</label>}
-          {!inline && (
-            <div className="validation-wrapper">
-              {validation &&
-                errors &&
-                errors.length > 0 &&
-                errors.map((el, key) => (
-                  <p className="error-message" key={key}>
-                    {el}
-                  </p>
-                ))}
-            </div>
-          )}
-        </div>
-
-        <div className="input-wrapper">
-          <div className="input-container">
-            <input
-              className={`${togglable ? "togglable" : ""}`}
-              name={name}
-              id={name}
-              type={togglable && revealed ? "text" : type}
-              required={required}
-              value={value}
-              placeholder={placeholder}
-              onBlur={(e) => {
-                if (typeof onFocusOut === "function") {
-                  onFocusOut({ name, value: e.target.value });
-                }
-              }}
-              onChange={(e) => {
-                if (typeof onChange === "function") {
-                  onChange({ name, value: e.target.value });
-                }
-              }}
-            />
-            {suffix && <span className="suffix">{suffix}</span>}
-            {togglable && (
-              <button
-                className="reveal-button"
-                type="button"
-                onClick={() => setRevealed(!revealed)}
-              >
-                {revealed ? (
-                  <FontAwesomeIcon icon={faEye} />
-                ) : (
-                  <FontAwesomeIcon icon={faEyeSlash} />
-                )}
-              </button>
-            )}
-          </div>
-          {inline && (
-            <div className="validation-wrapper">
-              {validation &&
-                errors &&
-                errors.length > 0 &&
-                errors.map((el, key) => (
-                  <p className="error-message" key={key}>
-                    {el}
-                  </p>
-                ))}
-            </div>
-          )}
-        </div>
-      </div>
-
-      {infos}
-    </div>
-  );
-};
-
-export default styled(TextField)(
-  ({ theme: { colors, fw } }) => css`
-    width: 100%;
-    .label-wrapper {
-      display: flex;
-      align-items: flex-end;
-      margin-bottom: 1rem;
-      label {
-        color: ${colors.formLabel};
-        font-weight: ${fw.bold};
-        font-size: 1.6rem;
-        display: block;
-      }
-      .validation-wrapper {
-        margin-left: 1.5rem;
-      }
-    }
-
-    .error-message {
-      font-size: 1.1rem;
-      font-weight: ${fw.semibold};
-      color: ${colors.formError};
-      margin: 0;
-      margin-top: 0.5rem;
-      line-height: 1.5;
-    }
-    input {
-      color: ${colors.formLabel};
-      border: 0.1rem solid ${colors.formBorder};
-      width: 100%;
-      font-size: 1.6rem;
-      border-radius: 1rem;
-      padding: 1rem;
-      transition: border-color 0.1s ease, box-shadow 0.2s ease;
-      box-shadow: 0 0 0 0 rgba(0, 0, 0, 1);
-      &.togglable {
-        padding-right: 5rem;
-      }
-      &:focus {
-        border-color: ${colors.primary};
-        box-shadow: 0 0 0.3rem 0 rgba(255, 191, 0, 1);
-        outline: none;
-      }
-    }
-    a {
-      font-size: 1.4rem;
-      color: ${colors.linkSecond};
-      margin-top: 0.75rem;
-      display: block;
-    }
-
-    &.error {
-      input {
-        border-color: ${colors.formError};
-        &:focus {
-          box-shadow: 0 0 0.3rem rgba(254, 124, 141, 1);
-        }
-      }
-    }
-
-    .input-wrapper,
-    .input-container {
-      position: relative;
-      width: 100%;
-    }
-
-    .reveal-button {
-      position: absolute;
-      right: 0;
-      top: 50%;
-      transform: translateY(-50%);
-      border: none;
-      outline: none;
-      background-color: transparent;
-      cursor: pointer;
-      font-size: 1.6rem;
-      color: ${colors.linkSecond};
-      margin-right: 1rem;
-    }
-
-    .suffix {
-      position: absolute;
-      right: 4rem;
-      top: 50%;
-      transform: translateY(-50%);
-      font-size: 1.3rem;
-      font-weight: 700;
-      color: ${colors.linkSecond};
-    }
-
-    .field-wrapper.inline {
-      display: flex;
-      .label-wrapper {
-        min-width: 17.5rem;
-        width: 17.5rem;
-        align-items: flex-start;
-        padding-top: 1rem;
-      }
-    }
-  `
-);
