--- client/src/components/Global/Form/Textarea.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Textarea = () => {
-  return <div></div>;
-};
-
-export default Textarea;
