+++ client/src/components/Scopes/Login/index.js
@@ -7,10 +7,10 @@ import { useUser } from "@/state/authentication";
 import { login } from "@/utils/api/authentication";
 import OAuth from "./OAuth";
 
-import Form from "@/components/Global/NewForm";
-import FieldWrapper from "@/components/Global/NewForm/FieldWrapper";
-import TextInput from "@/components/Global/NewForm/Fields/TextInput";
-import { useFields } from "@/components/Global/NewForm/utils";
+import Form from "@/components/Global/Form";
+import FieldWrapper from "@/components/Global/Form/FieldWrapper";
+import TextInput from "@/components/Global/Form/Fields/TextInput";
+import { useFields } from "@/components/Global/Form/utils";
 
 const LoginPage = ({ className }) => {
   const providers = [{ id: "google", name: "Google", icon: faGoogle }];
