+++ client/src/components/Scopes/Signup/index.js
@@ -4,11 +4,11 @@ import styled, { css } from "styled-components";
 
 import { useUser } from "@/state/authentication";
 import { signup } from "@/utils/api/authentication";
-import Form from "@/components/Global/NewForm";
-import FieldWrapper from "@/components/Global/NewForm/FieldWrapper";
-import TextInput from "@/components/Global/NewForm/Fields/TextInput";
-import PasswordConfirm from "@/components/Global/NewForm/Fields/PasswordConfirm";
-import { useFields } from "@/components/Global/NewForm/utils";
+import Form from "@/components/Global/Form";
+import FieldWrapper from "@/components/Global/Form/FieldWrapper";
+import TextInput from "@/components/Global/Form/Fields/TextInput";
+import PasswordConfirm from "@/components/Global/Form/Fields/PasswordConfirm";
+import { useFields } from "@/components/Global/Form/utils";
 
 const Signup = ({ className }) => {
   const [, authDispatch] = useUser();
